download 系のメソッドが、Post インスタンスを必要としないので Post の Extensions から Utils クラスに移動
